requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.1
gunicorn==21.2.0
requests==2.31.0
plyer==2.1.0
pyttsx3==2.90

